fix: show errors when decryption fails and remove duplication
