+++ src/renderer/components/Transaction/TransactionForm/TransactionFormDelegateRegistration.vue
@@ -95,6 +95,7 @@ import { ModalLoader } from '@/components/Modal'
 import { PassphraseInput } from '@/components/Passphrase'
 import TransactionService from '@/services/transaction'
 import WalletService from '@/services/wallet'
+import onSubmit from './mixin-on-submit'
 
 export default {
   name: 'TransactionFormDelegateRegistration',
@@ -109,6 +110,8 @@ export default {
     PassphraseInput
   },
 
+  mixins: [onSubmit],
+
   data: () => ({
     form: {
       fee: 0,
@@ -118,8 +121,7 @@ export default {
     },
     error: null,
     showEncryptLoader: false,
-    showLedgerLoader: false,
-    bip38Worker: null
+    showLedgerLoader: false
   }),
 
   computed: {
@@ -128,27 +130,7 @@ export default {
     }
   },
 
-  beforeDestroy () {
-    this.bip38Worker.send('quit')
-  },
-
   mounted () {
-    if (this.bip38Worker) {
-      this.bip38Worker.send('quit')
-    }
-    this.bip38Worker = this.$bgWorker.bip38()
-    this.bip38Worker.on('message', message => {
-      if (message.decodedWif === null) {
-        this.$error(this.$t('ENCRYPTION.FAILED_DECRYPT'))
-        this.showEncryptLoader = false
-      } else if (message.decodedWif) {
-        this.form.passphrase = null
-        this.form.wif = message.decodedWif
-        this.showEncryptLoader = false
-        this.submit()
-      }
-    })
-
     // Set default fees with v1 compatibility
     if (this.session_network.apiVersion === 1) {
       this.form.fee = V1.fees[this.$options.transactionType] / 1e8
@@ -160,19 +142,6 @@ export default {
   methods: {
     onFee (fee) {
       this.$set(this.form, 'fee', fee)
-    },
-
-    onSubmit () {
-      if (this.form.walletPassword && this.form.walletPassword.length) {
-        this.showEncryptLoader = true
-        this.bip38Worker.send({
-          bip38key: this.currentWallet.passphrase,
-          password: this.form.walletPassword,
-          wif: this.session_network.wif
-        })
-      } else {
-        this.submit()
-      }
     },
 
     async submit () {
