+++ src/renderer/components/Transaction/TransactionForm/TransactionFormVote.vue
@@ -148,6 +148,7 @@ import { ModalLoader } from '@/components/Modal'
 import { PassphraseInput } from '@/components/Passphrase'
 import WalletAddress from '@/components/Wallet/WalletAddress'
 import TransactionService from '@/services/transaction'
+import onSubmit from './mixin-on-submit'
 
 export default {
   name: 'TransactionFormVote',
@@ -164,6 +165,8 @@ export default {
     PassphraseInput,
     WalletAddress
   },
+
+  mixins: [onSubmit],
 
   props: {
     delegate: {
@@ -192,8 +195,7 @@ export default {
     forged: 0,
     voters: 0,
     showEncryptLoader: false,
-    showLedgerLoader: false,
-    bip38Worker: null
+    showLedgerLoader: false
   }),
 
   computed: {
@@ -239,28 +241,9 @@ export default {
     }
   },
 
-  beforeDestroy () {
-    this.bip38Worker.send('quit')
-  },
-
   mounted () {
     this.fetchForged()
     this.fetchVoters()
-    if (this.bip38Worker) {
-      this.bip38Worker.send('quit')
-    }
-    this.bip38Worker = this.$bgWorker.bip38()
-    this.bip38Worker.on('message', message => {
-      if (message.decodedWif === null) {
-        this.$error(this.$t('ENCRYPTION.FAILED_DECRYPT'))
-        this.showEncryptLoader = false
-      } else if (message.decodedWif) {
-        this.form.passphrase = null
-        this.form.wif = message.decodedWif
-        this.showEncryptLoader = false
-        this.submit()
-      }
-    })
 
     // Set default fees with v1 compatibility
     if (this.session_network.apiVersion === 1) {
@@ -286,19 +269,6 @@ export default {
 
     onFee (fee) {
       this.$set(this.form, 'fee', fee)
-    },
-
-    onSubmit () {
-      if (this.form.walletPassword && this.form.walletPassword.length) {
-        this.showEncryptLoader = true
-        this.bip38Worker.send({
-          bip38key: this.currentWallet.passphrase,
-          password: this.form.walletPassword,
-          wif: this.session_network.wif
-        })
-      } else {
-        this.submit()
-      }
     },
 
     async submit () {
